Rewrite punishment code to handle universal discipline settings

+++ admin/punishment/delete.php
@@ -34,18 +34,36 @@
 			$punishment = "{$row['DisciplineType']} on $dateinfo";
 			$log_pun    = "{$row['DisciplineType']} on {$row['Date']}";
 			
-			$query =	"SELECT discipline.WorkerUsername, discipline.RecordUsername, disciplineperms.Permissions " .
-						"       FROM discipline, disciplineperms " .
+			$query =	"SELECT ActiveTeacher FROM user WHERE Username='$username' AND ActiveTeacher=1";
+			$res =&  $db->query($query);
+			if(DB::isError($res)) die($res->getDebugInfo());           // Check for errors in query
+			if($row =& $res->fetchRow(DB_FETCHMODE_ASSOC)) {
+				$is_teacher = true;
+			} else {
+				$is_teacher = false;
+			}
+			
+			$query =    "SELECT Permissions FROM disciplineperms WHERE Username=\"$username\"";
+			$res =&  $db->query($query);
+			if(DB::isError($res)) die($res->getDebugInfo());           // Check for errors in query
+			if($row =& $res->fetchRow(DB_FETCHMODE_ASSOC)) {
+				$perm = $row['Permissions'];
+			} else {
+				$perm = $DEFAULT_PUN_PERM;
+			}
+	
+			$query =	"SELECT discipline.WorkerUsername, discipline.RecordUsername " .
+						"       FROM discipline " .
 						"WHERE  discipline.DisciplineIndex = $disciplineindex " .
-						"AND    disciplineperms.Username = \"$username\" " .
-						"AND    ((discipline.WorkerUsername = \"$username\" " .
-						"         OR discipline.RecordUsername = \"$username\") " .
-						"        AND disciplineperms.Permissions >= $PUN_PERM_MASS) " .
-						"OR     disciplineperms.Permissions >= $PUN_PERM_ALL ";
+						"AND    disciplineperms.Username = '$username' " .
+						"AND    ((discipline.WorkerUsername = '$username' " .
+						"         OR discipline.RecordUsername = '$username') " .
+						"        AND $perm >= $PUN_PERM_MASS) " .
+						"OR     $perm >= $PUN_PERM_ALL ";
 			$res =&  $db->query($query);
 			if(DB::isError($res)) die($res->getDebugInfo());           // Check for errors in query
 			/* Check whether current user is authorized to delete punishment */
-			if(dbfuncGetPermission($permissions, $PERM_ADMIN) or $res->numRows() > 0) {
+			if(dbfuncGetPermission($permissions, $PERM_ADMIN) or ($res->numRows() > 0 and $is_teacher)) {
 				$res =&  $db->query("DELETE FROM discipline " .          // Remove punishment from discipline table
 									"WHERE DisciplineIndex = $disciplineindex");
 				if(DB::isError($res)) die($res->getDebugInfo());          // Check for errors in query
